+++ download.go
@@ -12,8 +12,8 @@ import (
 
 func buildUrl(base, representationId, file string, partNum *int64) string {
 	if partNum != nil {
-		// $Number%05d$
 		file = strings.ReplaceAll(file, "$Number$", fmt.Sprintf("%05d", *partNum))
+		file = strings.ReplaceAll(file, "$Number%05d$", fmt.Sprintf("%05d", *partNum))
 	}
 	return base + strings.ReplaceAll(file, "$RepresentationID$", representationId)
 }
@@ -47,17 +47,23 @@ func downloadPart(url string) {
 }
 
 func getFilename(set *mpd.AdaptationSet) string {
+	if set == nil {
+		f, _ := os.CreateTemp("", "crdl-subs-*.ass")
+		return f.Name()
+	}
 	for _, representation := range set.Representations {
 		if representation.Height != nil {
-			return "temp_video.mp4"
+			f, _ := os.CreateTemp("", "crdl-video-*.mp4")
+			return f.Name()
 		} else if representation.Bandwidth != nil {
-			return "temp_audio.mp3"
+			f, _ := os.CreateTemp("", "crdl-audio-*.mp3")
+			return f.Name()
 		}
 	}
 	return ""
 }
 
-func downloadParts(baseUrl, representationId *string, set *mpd.AdaptationSet) {
+func downloadParts(baseUrl, representationId *string, set *mpd.AdaptationSet) string {
 	initUrl := buildUrl(*baseUrl, *representationId, *set.SegmentTemplate.Initialization, nil)
 	downloadPart(initUrl)
 
@@ -65,13 +71,14 @@ func downloadParts(baseUrl, representationId *string, set *mpd.AdaptationSet) {
 	for i, item := range timeline {
 		url := buildUrl(*baseUrl, *representationId, *set.SegmentTemplate.Media, &item)
 		downloadPart(url)
-		fmt.Printf("\rDownloaded %v of %v segments (%s)", i+1, len(timeline), humanSize(int64(len(parts))))
+		fmt.Printf("\rDownloaded %v of %v segments (%v%%)", i+1, len(timeline), (100*(i+1))/len(timeline))
 	}
 
 	fmt.Println("\nFinished downloading!")
 
 	// Write to a file
-	file, err := os.Create(getFilename(set))
+	filename := getFilename(set)
+	file, err := os.Create(filename)
 	if err != nil {
 		panic(err)
 	}
@@ -79,14 +86,16 @@ func downloadParts(baseUrl, representationId *string, set *mpd.AdaptationSet) {
 	if err != nil {
 		panic(err)
 	}
+	defer file.Close()
 	file.Write(decrypted)
-	file.Close()
 
 	// Empty parts
 	parts = []byte{}
+
+	return filename
 }
 
-func downloadSubs(url string) {
+func downloadSubs(url string) string {
 	req, err := http.NewRequest(http.MethodGet, url, nil)
 	if err != nil {
 		panic(err)
@@ -106,10 +115,93 @@ func downloadSubs(url string) {
 	}
 
 	// Write to a file
-	file, err := os.Create("subs.ass")
+	filename := getFilename(nil)
+	file, err := os.Create(filename)
 	if err != nil {
 		panic(err)
 	}
 	file.Write(body)
 	file.Close()
+
+	return filename
+}
+
+func downloadEpisode(contentId string, videoQuality, audioQuality, subtitlesLang *string, info EpisodeInfo) {
+	episode := getEpisode(contentId)
+	fmt.Printf("Downloading: %s (S%02vE%02v) from %s\n", info.Title, info.EpisodeMetadata.SeasonNumber, info.EpisodeMetadata.EpisodeNumber, info.EpisodeMetadata.SeriesTitle)
+
+	manifest := parseManifest(episode.ManifestURL)
+	pssh := getPssh(manifest)
+	if pssh == nil {
+		panic("PSSH not found")
+	}
+	videoSet := manifest.Period[0].AdaptationSets[0]
+	audioSet := manifest.Period[0].AdaptationSets[1]
+
+	// Get Widevine license
+	err := getLicense(*pssh, contentId, episode.Token)
+	if err != nil {
+		fmt.Printf("Error: %s", err)
+		os.Exit(1)
+	}
+
+	// Download subtitles
+	subtitles := episode.Subtitles[*subtitlesLang]
+	var subsFile string
+	if subtitles != nil {
+		fmt.Printf("Downloading subtitles for %s language...\n", languageNames[*subtitlesLang])
+		subsFile = downloadSubs(subtitles.URL)
+		fmt.Println("Downloaded subtitles!")
+	}
+
+	// Download video
+	baseUrl, representationId := getBaseUrl(videoSet, true, *videoQuality)
+	if baseUrl == nil {
+		print("Failed to get the video base URL, maybe the video quality you entered is wrong?\n")
+		os.Exit(1)
+	}
+	videoFile := downloadParts(baseUrl, representationId, videoSet)
+
+	// Download audio
+	audioBaseUrl, audioRepresentationId := getBaseUrl(audioSet, false, *audioQuality)
+	if audioBaseUrl == nil {
+		print("Failed to get the audio base URL, maybe the audio quality you entered is wrong?\n")
+		os.Exit(1)
+	}
+	audioFile := downloadParts(audioBaseUrl, audioRepresentationId, audioSet)
+
+	if success := deleteStream(contentId, episode.Token); !success {
+		print("Failed to remove the player stream, you will probably have issues downloading other episodes.\n")
+	}
+
+	renamed := strings.ReplaceAll(info.EpisodeMetadata.SeriesTitle, "'", "_")
+	renamed = strings.ReplaceAll(renamed, "/", "_")
+	renamed = strings.ReplaceAll(renamed, ":", "_")
+	if _, err := os.Stat(renamed); err != nil {
+		_ = os.MkdirAll(renamed, 0777)
+	}
+	outputFile := fmt.Sprintf("%s/%s S%02vE%02v [%s].mkv",
+		renamed, info.EpisodeMetadata.SeriesTitle, info.EpisodeMetadata.EpisodeNumber, info.EpisodeMetadata.SeasonNumber,
+		*videoQuality,
+	)
+	mergeEverything(videoFile, audioFile, subsFile, outputFile, subtitlesLang, info)
+}
+
+func downloadSeason(videoQuality, audioQuality, subtitlesLang *string, episodes []SeasonEpisode) {
+	fmt.Printf("Downloading season %v of %s (%v episodes)\n\n", episodes[0].SeasonNumber, episodes[0].SeriesTitle, len(episodes))
+
+	for _, episode := range episodes {
+		info := EpisodeInfo{
+			EpisodeMetadata: EpisodeMetadata{
+				SeriesTitle:        episode.SeriesTitle,
+				SeasonNumber:       episode.SeasonNumber,
+				EpisodeNumber:      episode.EpisodeNumber,
+				AudioLocale:        episode.AudioLocale,
+				Versions:           episode.Versions,
+				AvailabilityStarts: episode.AvailabilityStarts,
+			},
+			Title: episode.Title,
+		}
+		downloadEpisode(episode.ID, videoQuality, audioQuality, subtitlesLang, info)
+	}
 }