+++ mpd.go
@@ -40,9 +40,15 @@ func getBaseUrl(set *mpd.AdaptationSet, isVideoSet bool, quality string) (*strin
 				return &representation.BaseURL[0].Value, representation.ID
 			}
 		} else {
-			toInt, _ := strconv.ParseInt(strings.ReplaceAll(quality, "k", ""), 10, 64)
-			if (*representation.Bandwidth - 1) == uint64(toInt*1000) {
-				return &representation.BaseURL[0].Value, representation.ID
+			if strings.Contains(*representation.ID, "audio/") {
+				if strings.Contains(*representation.ID, quality) {
+					return &representation.BaseURL[0].Value, representation.ID
+				}
+			} else if representation.Bandwidth != nil {
+				toInt, _ := strconv.ParseInt(strings.ReplaceAll(quality, "k", ""), 10, 64)
+				if (*representation.Bandwidth - 1) == uint64(toInt*1000) {
+					return &representation.BaseURL[0].Value, representation.ID
+				}
 			}
 		}
 	}