+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.1.0
+
+- Added support for downloading entire seasons
+- Fixed MPD parsing
+- Temporary downloaded files (video, audio segments and subtitles) are now stored in the OS temporary files then deleted
+- Fixed FFmpeg merge command
+- Docs improvements
+- Support for `device_id.bin` and `private_key.pem` files
+
 ## 1.0.0
 
 Initial release